pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pymyio"
version = "0.1.0"
description = "Python bindings for the myIO d3.js chart library"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
authors = [{ name = "Morton Analytics" }]
keywords = ["d3", "visualization", "charts", "jupyter", "anywidget", "myio"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Framework :: Jupyter",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
  "anywidget>=0.10.0,<0.11",
  "traitlets>=5.9,<6",
  "ipywidgets>=8.0",
]

[project.optional-dependencies]
shiny = ["shinywidgets>=0.8.0", "shiny>=1.0"]
pandas = ["pandas>=1.5"]
polars = ["polars>=0.20"]
dev = [
  "pytest>=7",
  "pytest-playwright>=0.5",
  "pandas>=1.5",
  "shinywidgets>=0.8.0",
  "shiny>=1.0",
  "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/mortonanalytics/pymyIO"
Source = "https://github.com/mortonanalytics/pymyIO"
"Sister R package" = "https://github.com/mortonanalytics/myIO"

[tool.hatch.build.targets.wheel]
packages = ["src/pymyio"]
# Skip the symlinks under static/ — force-include below materializes them
# from the submodule so the wheel contains a single, canonical copy.
exclude = [
  "src/pymyio/static/myIOapi.js",
  "src/pymyio/static/style.css",
  "src/pymyio/static/lib/d3.min.js",
  "src/pymyio/static/lib/d3-hexbin.js",
  "src/pymyio/static/lib/d3-sankey.min.js",
]

# Materialize the engine into the wheel by following the symlinks under
# src/pymyio/static/ that point into the vendored myIO submodule. This way
# end-users who `pip install pymyio` get real files even though developers
# work from a single source of truth in vendor/myIO/.
[tool.hatch.build.targets.wheel.force-include]
"vendor/myIO/inst/htmlwidgets/myIO/myIOapi.js"     = "pymyio/static/myIOapi.js"
"vendor/myIO/inst/htmlwidgets/myIO/style.css"      = "pymyio/static/style.css"
"vendor/myIO/inst/htmlwidgets/lib/d3.min.js"       = "pymyio/static/lib/d3.min.js"
"vendor/myIO/inst/htmlwidgets/lib/d3-hexbin.js"    = "pymyio/static/lib/d3-hexbin.js"
"vendor/myIO/inst/htmlwidgets/lib/d3-sankey.min.js" = "pymyio/static/lib/d3-sankey.min.js"

# Sdist must include the submodule contents so wheels can be built from the
# tarball alone (PyPI doesn't fetch git submodules).
[tool.hatch.build.targets.sdist]
include = [
  "src/pymyio",
  "vendor/myIO/inst/htmlwidgets",
  "vendor/myIO/LICENSE",
  "tests",
  "examples",
  "README.md",
  "LICENSE",
  "pyproject.toml",
]

[tool.ruff]
line-length = 100
target-version = "py39"
